requirements.txt
python-telegram-bot==20.3
Pillow==9.5.0
python-dotenv==1.0.0
